requirements.txt
tweepy==4.15.0
APScheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
